refactor: reorganize the component structure on frontend-pwa
