+++ apps/frontend-pwa/src/components/common/OptionsDisplay.tsx
