+++ apps/frontend-pwa/src/components/home/CourseElement.tsx
