+++ apps/frontend-pwa/src/components/participant/Participant.tsx
