+++ apps/frontend-pwa/src/components/participant/Podium.tsx
