+++ apps/frontend-pwa/src/pages/course/[courseId]/docs/index.tsx
@@ -1,7 +1,7 @@
 import Markdown from '@klicker-uzh/markdown'
 import { H3 } from '@uzh-bf/design-system'
 import Image from 'next/image'
-import DocsLayout from '../../../../components/DocsLayout'
+import DocsLayout from '../../../../components/docs/DocsLayout'
 
 function Landing() {
   return (
