+++ apps/frontend-pwa/src/pages/course/[courseId]/docs/login.tsx
@@ -1,6 +1,6 @@
 import { H3 } from '@uzh-bf/design-system'
 import Image from 'next/image'
-import DocsLayout from '../../../../components/DocsLayout'
+import DocsLayout from '../../../../components/docs/DocsLayout'
 
 function Login() {
   return (
