+++ apps/frontend-pwa/src/pages/index.tsx
@@ -19,9 +19,9 @@ import dayjs from 'dayjs'
 import Link from 'next/link'
 import { useEffect, useMemo, useState } from 'react'
 import { twMerge } from 'tailwind-merge'
-import CourseElement from '../components/CourseElement'
+import CourseElement from '../components/home/CourseElement'
 import Layout from '../components/Layout'
-import SurveyPromotion from '../components/SurveyPromotion'
+import SurveyPromotion from '../components/surveys/SurveyPromotion'
 import {
   determineInitialSubscriptionState,
   subscribeParticipant,
@@ -72,8 +72,8 @@ const Index = function () {
         courses: [
           ...acc.courses,
           {
-            id: participation.course.id,
-            displayName: participation.course.displayName,
+            id: participation.course?.id,
+            displayName: participation.course?.displayName,
             isSubscribed:
               participation.subscriptions &&
               participation.subscriptions.length > 0,
